+++ src/api/endpoints/orders/get-received-bids/v1.ts
@@ -13,7 +13,7 @@ import _ from "lodash";
 const version = "v1";
 
 export const getReceivedBidsV1Options: RouteOptions = {
-  description: "Received Bids (offers)",
+  description: "User Bids Received",
   notes:
     "Get a list of received bids (offers), filtered by address. This API is designed for efficiently receiving all bids, for real-time ingestion.",
   tags: ["api", "Orders"],