+++ src/api/endpoints/tokens/get-tokens-ids/v1.ts
@@ -10,7 +10,7 @@ import { regex, toBuffer } from "@/common/utils";
 const version = "v1";
 
 export const getTokensIdsV4Options: RouteOptions = {
-  description: "All token ids a collection / contract / token set",
+  description: "Token IDs",
   notes:
     "This API is optimized for quickly fetching a list of tokens ids in by collection, contract, token set id. ",
   tags: ["api", "Tokens"],